Move sub-classes of "KeyboardToolTipStateMachine", "ListItemConverter", "LayoutUtils" and "ListViewSubItem" classes to their own files
