+++ src/System.Windows.Forms/src/System/Windows/Forms/Layout/LayoutUtils.MeasureTextCache.PreferredSizeCache.cs
@@ -1,0 +1,23 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Drawing;
+
+namespace System.Windows.Forms.Layout
+{
+    internal partial class LayoutUtils
+    {
+        private struct PreferredSizeCache
+        {
+            public PreferredSizeCache(Size constrainingSize, Size preferredSize)
+            {
+                ConstrainingSize = constrainingSize;
+                PreferredSize = preferredSize;
+            }
+
+            public Size ConstrainingSize;
+            public Size PreferredSize;
+        }
+    }
+}
