+++ src/System.Windows.Forms/src/System/Windows/Forms/Layout/LayoutUtils.cs
@@ -10,7 +10,7 @@ namespace System.Windows.Forms.Layout
 {
     // Utilities used by layout code.  If you use these outside of the layout
     // namespace, you should probably move them to WindowsFormsUtils.
-    internal class LayoutUtils
+    internal partial class LayoutUtils
     {
         public static readonly Size MaxSize = new Size(int.MaxValue, int.MaxValue);
         public static readonly Size InvalidSize = new Size(int.MinValue, int.MinValue);
@@ -609,253 +609,6 @@ namespace System.Windows.Forms.Layout
         {
             bounds.X = withinBounds.Width - bounds.Right;
             return bounds;
-        }
-
-        ///  MeasureTextCache
-        ///  3000 character strings take 9 seconds to load the form
-        public sealed class MeasureTextCache
-        {
-            private Size unconstrainedPreferredSize = LayoutUtils.InvalidSize;
-            private const int MaxCacheSize = 6;           // the number of preferred sizes to store
-            private int nextCacheEntry = -1;              // the next place in the ring buffer to store a preferred size
-
-            private PreferredSizeCache[]? sizeCacheList;   // MRU of size MaxCacheSize
-
-            ///  InvalidateCache
-            ///  Clears out the cached values, should be called whenever Text, Font or a TextFormatFlag has changed
-            public void InvalidateCache()
-            {
-                unconstrainedPreferredSize = LayoutUtils.InvalidSize;
-                sizeCacheList = null;
-            }
-
-            ///  GetTextSize
-            ///  Given constraints, format flags a font and text, determine the size of the string
-            ///  employs an MRU of the last several constraints passed in via a ring-buffer of size MaxCacheSize.
-            ///  Assumes Text and TextFormatFlags are the same, if either were to change, a call to
-            ///  InvalidateCache should be made
-            public Size GetTextSize(string? text, Font? font, Size proposedConstraints, TextFormatFlags flags)
-            {
-                if (!TextRequiresWordBreak(text, font, proposedConstraints, flags))
-                {
-                    // Text fits within proposed width
-
-                    // IF we're here, this means we've got text that can fit into the proposedConstraints
-                    // without wrapping.  We've determined this because our
-
-                    // as a side effect of calling TextRequiresWordBreak,
-                    // unconstrainedPreferredSize is set.
-                    return unconstrainedPreferredSize;
-                }
-                else
-                {
-                    // Text does NOT fit within proposed width - requires WordBreak
-
-                    // IF we're here, this means that the wrapping width is smaller
-                    // than our max width.  For example: we measure the text with infinite
-                    // bounding box and we determine the width to fit all the characters
-                    // to be 200 px wide.  We would come here only for proposed widths less
-                    // than 200 px.
-
-                    // Create our ring buffer if we dont have one
-                    if (sizeCacheList is null)
-                    {
-                        sizeCacheList = new PreferredSizeCache[MaxCacheSize];
-                    }
-
-                    // check the existing constraints from previous calls
-                    foreach (PreferredSizeCache sizeCache in sizeCacheList)
-                    {
-                        if (sizeCache.ConstrainingSize == proposedConstraints)
-                        {
-                            return sizeCache.PreferredSize;
-                        }
-                        else if ((sizeCache.ConstrainingSize.Width == proposedConstraints.Width)
-                                  && (sizeCache.PreferredSize.Height <= proposedConstraints.Height))
-                        {
-                            // Caching a common case where the width matches perfectly, and the stored preferred height
-                            // is smaller or equal to the constraining size.
-                            //        prefSize = GetPreferredSize(w,Int32.MaxValue);
-                            //        prefSize = GetPreferredSize(w,prefSize.Height);
-
-                            return sizeCache.PreferredSize;
-                        }
-
-                        //
-                    }
-
-                    // if we've gotten here, it means we dont have a cache entry, therefore
-                    // we should add a new one in the next available slot.
-                    Size prefSize = TextRenderer.MeasureText(text, font, proposedConstraints, flags);
-                    nextCacheEntry = (nextCacheEntry + 1) % MaxCacheSize;
-                    sizeCacheList[nextCacheEntry] = new PreferredSizeCache(proposedConstraints, prefSize);
-
-                    return prefSize;
-                }
-            }
-
-            ///  GetUnconstrainedSize
-            ///  Gets the unconstrained (Int32.MaxValue, Int32.MaxValue) size for a piece of text
-            private Size GetUnconstrainedSize(string? text, Font? font, TextFormatFlags flags)
-            {
-                if (unconstrainedPreferredSize == InvalidSize)
-                {
-                    // we also investigated setting the SingleLine flag, however this did not yield as much benefit as the word break
-                    // and had possibility of causing internationalization issues.
-
-                    flags = (flags & ~TextFormatFlags.WordBreak); // rip out the wordbreak flag
-                    unconstrainedPreferredSize = TextRenderer.MeasureText(text, font, MaxSize, flags);
-                }
-
-                return unconstrainedPreferredSize;
-            }
-
-            ///  TextRequiresWordBreak
-            ///  If you give the text all the space in the world it wants, then there should be no reason
-            ///  for it to break on a word.  So we find out what the unconstrained size is (Int32.MaxValue, Int32.MaxValue)
-            ///  for a string - eg. 35, 13.  If the size passed in has a larger width than 35, then we know that
-            ///  the WordBreak flag is not necessary.
-            public bool TextRequiresWordBreak(string? text, Font? font, Size size, TextFormatFlags flags)
-            {
-                // if the unconstrained size of the string is larger than the proposed width
-                // we need the word break flag, otherwise we dont, its a perf hit to use it.
-                return GetUnconstrainedSize(text, font, flags).Width > size.Width;
-            }
-
-            private struct PreferredSizeCache
-            {
-                public PreferredSizeCache(Size constrainingSize, Size preferredSize)
-                {
-                    ConstrainingSize = constrainingSize;
-                    PreferredSize = preferredSize;
-                }
-
-                public Size ConstrainingSize;
-                public Size PreferredSize;
-            }
-        }
-    }
-
-    // Frequently when you need to do a PreformLayout, you also need to invalidate the
-    // PreferredSizeCache (you are laying out because you know that the action has changed
-    // the PreferredSize of the control and/or its container).  LayoutTransaction wraps both
-    // of these operations into one, plus adds a check for null to make our code more
-    // concise.
-    //
-    // Usage1: (When we are not calling to other code which may cause a layout:)
-    //
-    //  LayoutTransaction.DoLayout(ParentInternal, this, PropertyNames.Bounds);
-    //
-    // Usage2: (When we need to wrap code which may cause additional layouts:)
-    //
-    //  using(new LayoutTransaction(ParentInternal, this, PropertyNames.Bounds)) {
-    //      OnBoundsChanged();
-    //  }
-    //
-    // The second usage spins off 12b for garbage collection, but I did some profiling and
-    // it didn't seem significant (we were spinning off more from LayoutEventArgs.)
-    internal sealed class LayoutTransaction : IDisposable
-    {
-        readonly Control? _controlToLayout;
-        readonly bool _resumeLayout;
-
-#if DEBUG
-        readonly int _layoutSuspendCount;
-#endif
-        public LayoutTransaction(Control? controlToLayout, IArrangedElement? controlCausingLayout, string? property) :
-            this(controlToLayout, controlCausingLayout, property, true)
-        {
-        }
-
-        public LayoutTransaction(Control? controlToLayout, IArrangedElement? controlCausingLayout, string? property, bool resumeLayout)
-        {
-            CommonProperties.xClearPreferredSizeCache(controlCausingLayout);
-            _controlToLayout = controlToLayout;
-
-            _resumeLayout = resumeLayout;
-            if (_controlToLayout != null)
-            {
-#if DEBUG
-                _layoutSuspendCount = _controlToLayout.LayoutSuspendCount;
-#endif
-                _controlToLayout.SuspendLayout();
-                CommonProperties.xClearPreferredSizeCache(_controlToLayout);
-
-                // Same effect as calling performLayout on Dispose but then we would have to keep
-                // controlCausingLayout and property around as state.
-                if (resumeLayout)
-                {
-                    _controlToLayout.PerformLayout(new LayoutEventArgs(controlCausingLayout, property));
-                }
-            }
-        }
-
-        public void Dispose()
-        {
-            if (_controlToLayout != null)
-            {
-                _controlToLayout.ResumeLayout(_resumeLayout);
-
-#if DEBUG
-                Debug.Assert(_controlToLayout.LayoutSuspendCount == _layoutSuspendCount, "Suspend/Resume layout mismatch!");
-#endif
-            }
-        }
-
-        // This overload should be used when a property has changed that affects preferred size,
-        // but you only want to layout if a certain condition exists (say you want to layout your
-        // parent because your preferred size has changed).
-        public static IDisposable CreateTransactionIf(bool condition, Control? controlToLayout, IArrangedElement? elementCausingLayout, string? property)
-        {
-            if (condition)
-            {
-                return new LayoutTransaction(controlToLayout, elementCausingLayout, property);
-            }
-            else
-            {
-                CommonProperties.xClearPreferredSizeCache(elementCausingLayout);
-                return new NullLayoutTransaction();
-            }
-        }
-
-        public static void DoLayout(IArrangedElement? elementToLayout, IArrangedElement? elementCausingLayout, string? property)
-        {
-            if (elementCausingLayout != null)
-            {
-                CommonProperties.xClearPreferredSizeCache(elementCausingLayout);
-                if (elementToLayout != null)
-                {
-                    CommonProperties.xClearPreferredSizeCache(elementToLayout);
-                    elementToLayout.PerformLayout(elementCausingLayout, property);
-                }
-            }
-
-            Debug.Assert(elementCausingLayout != null, "LayoutTransaction.DoLayout - elementCausingLayout is null, no layout performed - did you mix up your parameters?");
-        }
-
-        // This overload should be used when a property has changed that affects preferred size,
-        // but you only want to layout if a certain condition exists (say you want to layout your
-        // parent because your preferred size has changed).
-        public static void DoLayoutIf(bool condition, IArrangedElement? elementToLayout, IArrangedElement? elementCausingLayout, string? property)
-        {
-            if (!condition)
-            {
-                if (elementCausingLayout != null)
-                {
-                    CommonProperties.xClearPreferredSizeCache(elementCausingLayout);
-                }
-            }
-            else
-            {
-                LayoutTransaction.DoLayout(elementToLayout, elementCausingLayout, property);
-            }
-        }
-    }
-
-    internal struct NullLayoutTransaction : IDisposable
-    {
-        public void Dispose()
-        {
         }
     }
 }
