+++ src/System.Windows.Forms/src/System/Windows/Forms/ListItemConverter.cs
@@ -186,59 +186,4 @@ namespace System.Windows.Forms
             return base.ConvertTo(context, culture, value, destinationType);
         }
     }
-
-    internal class ListViewSubItemConverter : ExpandableObjectConverter
-    {
-        public override bool CanConvertTo(ITypeDescriptorContext context, Type destinationType)
-        {
-            if (destinationType == typeof(InstanceDescriptor))
-            {
-                return true;
-            }
-
-            return base.CanConvertTo(context, destinationType);
-        }
-
-        public override object ConvertTo(ITypeDescriptorContext context, CultureInfo culture, object value, Type destinationType)
-        {
-            if (destinationType is null)
-            {
-                throw new ArgumentNullException(nameof(destinationType));
-            }
-
-            if (destinationType == typeof(InstanceDescriptor) && value is ListViewItem.ListViewSubItem item)
-            {
-                ConstructorInfo ctor;
-
-                // Subitem has custom style
-                if (item.CustomStyle)
-                {
-                    ctor = typeof(ListViewItem.ListViewSubItem).GetConstructor(new Type[]
-                    {
-                        typeof(ListViewItem),
-                        typeof(string),
-                        typeof(Color),
-                        typeof(Color),
-                        typeof(Font)
-                    });
-                    Debug.Assert(ctor != null, "Expected the constructor to exist.");
-                    return new InstanceDescriptor(ctor, new object[]
-                    {
-                        null,
-                        item.Text,
-                        item.ForeColor,
-                        item.BackColor,
-                        item.Font
-                    }, true);
-                }
-
-                // Otherwise, just use the text constructor
-                ctor = typeof(ListViewItem.ListViewSubItem).GetConstructor(new Type[] { typeof(ListViewItem), typeof(string) });
-                Debug.Assert(ctor != null, "Expected the constructor to exist.");
-                return new InstanceDescriptor(ctor, new object[] { null, item.Text }, true);
-            }
-
-            return base.ConvertTo(context, culture, value, destinationType);
-        }
-    }
 }
